채팅: 유저, 채팅봇 메세지 컴포넌트 생성
