+++ src/pages/Chat/chat.module.css
@@ -33,7 +33,8 @@
 
 .chatSection {
   height: 400px;
-  overflow: scroll;
+  overflow-y: scroll;
+  overflow-x: hidden;
   padding: 10px;
 }
 
@@ -44,7 +45,12 @@
 .chatbotSection {
   display: flex;
   text-align: left;
-  margin-bottom: 5px;
+  margin-top: 5px;
+}
+.bubbleSection {
+  display: flex;
+  align-items: center;
+  text-align: left;
 }
 
 .chatbotMsg {
@@ -52,7 +58,19 @@
   flex-direction: column;
   align-items: flex-start;
   margin-left: 1rem;
-  margin-top: 5px;
+  margin-top: 10px;
+}
+.bubbleImg {
+  visibility: hidden;
+}
+.bubbleList {
+  margin-left: 1rem;
+}
+.bookChat {
+  position: relative;
+  left: 77px;
+  width: 100%;
+  margin-bottom: 10px;
 }
 
 .msg {
@@ -81,6 +99,7 @@
   background-color: makerYellow;
   color: #fff;
   margin-right: 0.5rem;
+  margin-top: 5px;
 }
 
 .chatbotName {
