+++ src/pages/home/home.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import styles from "./home.module.css";
 import { useSelector } from "react-redux";
-import Chat from "./chat";
+import Chat from "../Chat/chat";
 
 const Home = (props) => {
   const { user: currentUser } = useSelector((state) => state.auth);
